requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
plotly==5.15.0
prophet==1.1.4
openpyxl==3.1.2
xlsxwriter==3.1.2
python-dateutil==2.8.2
pytz==2023.3
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.6
python-dotenv>=1.0.0
gunicorn>=20.1.0
seaborn>=0.11.2
sqlalchemy>=2.0.0
pystan>=2.19.1.1
holidays>=0.25
lunardate>=0.0.9
convertdate>=2.4.0
cmdstanpy>=1.1.0
pandas-datareader>=0.10.0
